chore(2020): add comments to hooks
