+++ packages/twenty-twenty-theme/src/components/mobile/menu-modal.js
@@ -1,10 +1,10 @@
 import React, { useRef } from "react";
 import { styled, connect, Global } from "frontity";
-import Link from "./link";
-import { CloseNavToggle } from "./nav-toggle";
-import { CloseIcon } from "./icons";
-import useFocusTrap from "./use-trap-focus";
-import useFocusEffect from "./use-focus-effect";
+import Link from "../link";
+import { CloseNavToggle } from "../nav-toggle";
+import { CloseIcon } from "../icons";
+import useFocusTrap from "../hooks/use-trap-focus";
+import useFocusEffect from "../hooks/use-focus-effect";
 
 const MobileMenuModal = ({ state, actions }) => {
   const { menu, isMobileMenuOpen } = state.theme;
