+++ packages/twenty-twenty-theme/src/components/mobile/search-button.js
@@ -1,7 +1,7 @@
 import { connect } from "frontity";
 import React from "react";
-import { SearchIcon } from "./icons";
-import { LabeledIcon, SearchToggle, ToggleWrapper } from "./nav-toggle";
+import { SearchIcon } from "../icons";
+import { LabeledIcon, SearchToggle, ToggleWrapper } from "../nav-toggle";
 
 const MobileSearchButton = ({ state, actions }) => {
   // Get the state of the search modal
