--- packages/twenty-twenty-theme/src/components/use-trap-focus.js
@@ -1,40 +1,0 @@
-import { useEffect } from "react";
-
-function trapFocus(element) {
-  var focusableEls = element.querySelectorAll(
-      "a[href]:not([disabled]), button:not([disabled]), textarea:not([disabled]), input:not([disabled]), select:not([disabled])"
-    ),
-    firstFocusableEl = focusableEls[0];
-  const lastFocusableEl = focusableEls[focusableEls.length - 1];
-  const KEYCODE_TAB = 9;
-
-  element.addEventListener("keydown", function(e) {
-    var isTabPressed = e.key === "Tab" || e.keyCode === KEYCODE_TAB;
-
-    if (!isTabPressed) {
-      return;
-    }
-
-    if (e.shiftKey) {
-      /* shift + tab */ if (document.activeElement === firstFocusableEl) {
-        lastFocusableEl.focus();
-        e.preventDefault();
-      }
-    } /* tab */ else {
-      if (document.activeElement === lastFocusableEl) {
-        firstFocusableEl.focus();
-        e.preventDefault();
-      }
-    }
-  });
-}
-
-function useFocusTrap(ref, isOpen) {
-  useEffect(() => {
-    if (isOpen && ref.current) {
-      trapFocus(ref.current);
-    }
-  }, [isOpen]);
-}
-
-export default useFocusTrap;
